+++ src/components/common/Parts/PaginationWrapper/index.tsx
@@ -1,0 +1,1 @@
+export { PaginationWrapper } from './PaginationWrapper';
